+++ src/game/nlcore/common/interface.ts
@@ -1,1 +1,2 @@
 export {Gallery} from "../elements/built-in/Gallery";
+export {DevTools} from "../elements/built-in/DevTools";